Add test for avg win percentage by opponent
